1) Fractional digits rounded to two. 2) Invoice design corrected. 3) Customer drop down menu added to invoice

+++ OrderManagementSystem.Entity/Models/OrderDetail.cs
@@ -15,6 +15,6 @@ public partial class OrderDetail
     public DateTime? OrderDate { get; set; }
     public decimal? UnitPrice { get; set; }
     public int? Quantity { get; set; }
-    public decimal? ItemTotal { get; set; } 
+    public float? ItemTotal { get; set; } 
 
 }
